Introduce lift module
Introduce a `pollicy::lift` module and move the lifting types and
logic there. Add deprecated re-exports although these may not
function correctly, I didn't check.
